Extract MultipartFormData to it’s own file.
